DayTradingProject/requirements.txt
ccxt
python-dotenv
pandas
pyarrow
ta
backtesting
scikit-learn
joblib
loguru
typer
tenacity
pydantic
requests